переносит enable-, disableCollection и enableForm в util.js
