+++ js/util.js
@@ -8,8 +8,33 @@ const isEscapeEvent = (evt) => {
   return (evt.keyCode === 27);
 };
 
-// const isEnterEvent = (evt) => {
-//   return (evt.keyCode === 13) || (evt.keyCode === 32);
-// };
+const enableCollection = (collection) => {
+  collection.forEach((element) => {
+    element.removeAttribute('disabled');
+  });
+};
 
-export { createElementFromTemplate, isEscapeEvent }
+const disableCollection = (collection) => {
+  collection.forEach((element) => {
+    element.setAttribute('disabled', '');
+  });
+};
+
+const enableForm = (className, enableFlag) => {
+  const formElement = document.querySelector(`.${className}`);
+  const disabledFormClass = `${className}--disabled`;
+  const selectors = ['fieldset', 'input', 'select'];
+  if (enableFlag) {
+    formElement.classList.remove(disabledFormClass);
+    selectors.forEach((selector) => {
+      enableCollection(formElement.querySelectorAll(selector));
+    });
+  } else {
+    formElement.classList.add(disabledFormClass);
+    selectors.forEach((selector) => {
+      disableCollection(formElement.querySelectorAll(selector));
+    });
+  }
+}
+
+export { createElementFromTemplate, isEscapeEvent, enableCollection, disableCollection, enableForm }
